Simplify cache tag generation by centralizing `getTagsToRevalidateByEntity`. Remove redundant `getNewsArticleCacheTags` logic and update usages accordingly.
